Refactor themes structure
