+++ src/themes/_base/_theme.scss
@@ -1,4 +1,4 @@
-@use '../styles/mixins' as *;
+@use '../../styles/mixins' as *;
 @use './palette' as palette;
 
 // Theme options
