+++ src/themes/dark/theme.scss
@@ -1,5 +1,5 @@
 /* Import the base theme */
-@use '../base';
+@use '../_base/theme';
 
 /* Card background color variants */
 .defaultCardBackground1 {
