+++ src/utils/reactUtils.tsx
@@ -6,7 +6,7 @@ import { createRoot } from 'react-dom/client';
 import { ApiProvider } from 'hooks/useApi';
 import { UserSettingsProvider } from 'hooks/useUserSettings';
 import { WebConfigProvider } from 'hooks/useWebConfig';
-import appTheme from 'themes/themes';
+import appTheme from 'themes';
 import { queryClient } from 'utils/query/queryClient';
 
 export const renderComponent = <P extends object> (
